+++ web/apps/photos/src/services/deduplicationService.ts
@@ -1,3 +1,4 @@
+import { hasFileHash } from "@/media/file";
 import { FILE_TYPE } from "@/media/file-type";
 import type { Metadata } from "@/media/types/file";
 import log from "@/next/log";
@@ -5,7 +6,6 @@ import HTTPService from "@ente/shared/network/HTTPService";
 import { getEndpoint } from "@ente/shared/network/api";
 import { getToken } from "@ente/shared/storage/localStorage/helpers";
 import { EnteFile } from "types/file";
-import { hasFileHash } from "utils/upload";
 
 const ENDPOINT = getEndpoint();
