--- web/apps/photos/src/utils/upload/uploadRetrier.ts
@@ -1,29 +1,0 @@
-import { wait } from "@ente/shared/utils";
-
-const retrySleepTimeInMilliSeconds = [2000, 5000, 10000];
-
-export async function retryHTTPCall(
-    func: () => Promise<any>,
-    checkForBreakingError?: (error) => void,
-): Promise<any> {
-    const retrier = async (
-        func: () => Promise<any>,
-        attemptNumber: number = 0,
-    ) => {
-        try {
-            const resp = await func();
-            return resp;
-        } catch (e) {
-            if (checkForBreakingError) {
-                checkForBreakingError(e);
-            }
-            if (attemptNumber < retrySleepTimeInMilliSeconds.length) {
-                await wait(retrySleepTimeInMilliSeconds[attemptNumber]);
-                return await retrier(func, attemptNumber + 1);
-            } else {
-                throw e;
-            }
-        }
-    };
-    return await retrier(func);
-}
