+++ web/packages/media/file.ts
@@ -1,0 +1,4 @@
+import type { Metadata } from "./types/file";
+
+export const hasFileHash = (file: Metadata) =>
+    !!file.hash || (!!file.imageHash && !!file.videoHash);
